Removing tests inside ECFactory to their corresponding elliptic curve type, refactoring test cases to use ECPrivateKey.fromBase58ToPrivateKey
