+++ src/main/scala/org/bitcoins/core/crypto/ECFactory.scala
@@ -116,20 +116,6 @@ trait ECFactory extends Factory[BaseECKey] {
   override def fromBytes(bytes : Seq[Byte]) : BaseECKey = privateKey(bytes)
 
 
-  /**
-   * Takes in a base58 string and converts it into a private key
-   * @param base58
-   * @return
-   */
-  def fromBase58ToPrivateKey(base58 : String) : ECPrivateKey = {
-    val decodedBase58 : Seq[Byte] = Base58.decode(base58)
-    //Drop(1) will drop the network byte. The last 5 bytes are dropped included the checksum (4 bytes), and 0x01 byte that
-    //is appended to compressed keys (which we implemented as the default option).
-    val trim = decodedBase58.drop(1).dropRight(5)
-    val privateKeyBytesToHex = BitcoinSUtil.encodeHex(trim)
-    ECFactory.privateKey(privateKeyBytesToHex)
-  }
-
 }
 
 object ECFactory extends ECFactory
