+++ src/test/scala/org/bitcoins/core/util/CryptoTestUtil.scala
@@ -2,7 +2,7 @@ package org.bitcoins.core.util
 
 import org.bitcoinj.core.DumpedPrivateKey
 import org.bitcoins.core.config.TestNet3
-import org.bitcoins.core.crypto.ECFactory
+import org.bitcoins.core.crypto.{ECFactory, ECPrivateKey}
 
 /**
  * Created by chris on 3/7/16.
@@ -13,7 +13,7 @@ trait CryptoTestUtil {
   def privateKeyHex = BitcoinSUtil.encodeHex(privateKeyBytes)
   def bitcoinjDumpedPrivateKey = new DumpedPrivateKey(BitcoinJTestUtil.params,privateKeyBase58)
   def bitcoinjPrivateKey = bitcoinjDumpedPrivateKey.getKey
-  def privateKey = ECFactory.fromBase58ToPrivateKey(privateKeyBase58)
+  def privateKey = ECPrivateKey.fromBase58ToPrivateKey(privateKeyBase58)
 
 }
